requirements.txt
Django>2.0
psycopg2-binary
django-import-export